ING-842: Fixed conflicted transactions state bits.
This commit also cleans up the overall handling of our state bits
to try and make it less likely that we encounter a similar issue
in the future.

+++ transactionsx/transactionattempt.go
@@ -84,7 +84,7 @@ func (t *TransactionAttempt) UpdateState(opts TransactionUpdateStateOptions) {
 	if opts.Reason == TransactionErrorReasonTransactionExpired {
 		stateBits |= transactionStateBitHasExpired
 	}
-	t.applyStateBits(stateBits, uint32(opts.Reason))
+	t.applyStateBits(stateBits, opts.Reason)
 
 	t.lock.Lock()
 	if opts.State > 0 {
